pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "eth-account>=0.13.6",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-jwt-extended>=4.7.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "routes>=2.5.1",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "requests>=2.32.3",
    "web3>=7.10.0",
    "flask-wtf>=1.2.2",
    "sendgrid>=6.11.0",
    "wtforms>=3.2.1",
    "xrpl-py>=4.1.0",
    "stripe>=12.0.0",
    "paypalrestsdk>=1.13.3",
    "pdfkit>=1.0.0",
    "weasyprint>=65.1",
    "trafilatura>=2.0.0",
    "markdown>=3.8",
    "pycryptodome>=3.22.0",
    "paramiko>=3.5.1",
    "pysftp>=0.2.9",
    "bots-ediint>=3.3.1",
    "crypto>=1.4.1",
    "openai>=1.76.0",
    "beautifulsoup4>=4.13.4",
    "pyppeteer>=0.0.25",
    "fpdf>=1.7.2",
    "python-docx>=1.1.2",
    "psutil>=7.0.0",
    "docx>=0.2.4",
    "jinja2>=3.1.6",
    "qrcode>=8.2",
    "python-dotenv>=1.1.0",
]
